edk2-ovmf: Fix TPM support
Summary: - ovmf 202202 renmaed the TPM_ENABLE build flag for enabling TPM1/2 support in the firmware to TPM1_ENABLE and TPM2_ENABLE (for TPM v1 and TPM v2 respectively). If the new flags were not used TPM support was not enabled in the firmware which would break VMs that had a software/passthrough TPM.

Test Plan: - Booted Win 11 VM and confirmed that TPM was working.

Reviewers: #triage_team, Staudey

Reviewed By: Staudey

Subscribers: Staudey

Differential Revision: https://dev.getsol.us/D13325